requirements.txt
huggingface_hub==0.22.2
accelerate
diffusers
invisible_watermark
torch
transformers
xformers
torchvision
gradio
Pillow
detoxify
altair==4.1.0
streamlit
